Extract date range methods into a trait
