+++ includes/class-sensei-autoloader.php
@@ -74,6 +74,7 @@ class Sensei_Autoloader {
 			new Sensei_Autoloader_Bundle( 'course-theme' ),
 			new Sensei_Autoloader_Bundle( 'course-video' ),
 			new Sensei_Autoloader_Bundle( 'course-video/blocks' ),
+			new Sensei_Autoloader_Bundle( 'reports/helper' ),
 			new Sensei_Autoloader_Bundle( 'reports/overview/data-provider' ),
 			new Sensei_Autoloader_Bundle( 'reports/overview/list-table' ),
 		);
